Moves state package_test.go into the state package.
This will allow further testing flexibility in future.
