+++ state/no_skip_test.go
@@ -3,7 +3,7 @@
 
 // +build !skip_state_tests
 
-package state_test
+package state
 
 // runStateTests controls whether to run the state tests - in this
 // case the skip_state_tests build tag hasn't been set so they'll be
